fix(web): remove waitlist screens
