+++ apps/web/src/app/page.tsx
@@ -1,5 +1,5 @@
 import { headers } from 'next/headers';
-import { ComingSoon } from '@/components/shared/ComingSoon';
+import { LandingPage } from '@/components/landing';
 import { isWaitlistHostname } from '@/lib/host-routing';
 import { HomePageClient } from './HomePageClient';
 
@@ -9,7 +9,7 @@ export default async function HomePage() {
   const isWaitlistHost = isWaitlistHostname(hostname);
 
   if (isWaitlistHost) {
-    return <ComingSoon />;
+    return <LandingPage />;
   }
 
   return <HomePageClient />;
